refactor: add import/no-internal-modules ESLint rule (#7010)
* refactor: add import/no-internal-modules ESLint rule
