+++ apps/cowswap-frontend/src/common/pure/ReceiveAmount/index.tsx
@@ -9,7 +9,7 @@ import { BalanceAndSubsidy } from 'legacy/hooks/useCowBalanceAndSubsidy'
 
 import { getOrderTypeReceiveAmounts } from 'modules/trade'
 import { useEstimatedBridgeBuyAmount } from 'modules/trade'
-import { ReceiveAmountInfo } from 'modules/trade/types'
+import { ReceiveAmountInfo } from 'modules/trade'
 
 import * as styledEl from './styled'
