+++ apps/cowswap-frontend/src/modules/advancedOrders/state/advancedOrdersAtom.ts
@@ -4,8 +4,12 @@ import { atomWithStorage } from 'jotai/utils'
 import { getJotaiIsolatedStorage } from '@cowprotocol/core'
 import { OrderKind, SupportedChainId } from '@cowprotocol/cow-sdk'
 
-import { DEFAULT_TRADE_DERIVED_STATE, TradeDerivedState } from 'modules/trade/types/TradeDerivedState'
-import { ExtendedTradeRawState, getDefaultTradeRawState } from 'modules/trade/types/TradeRawState'
+import {
+  DEFAULT_TRADE_DERIVED_STATE,
+  ExtendedTradeRawState,
+  getDefaultTradeRawState,
+  TradeDerivedState,
+} from 'modules/trade'
 
 export interface AdvancedOrdersDerivedState extends TradeDerivedState {
   readonly isUnlocked: boolean
