+++ apps/cowswap-frontend/src/modules/limitOrders/state/limitOrdersRawStateAtom.ts
@@ -6,11 +6,15 @@ import { getJotaiIsolatedStorage } from '@cowprotocol/core'
 import { OrderKind, SupportedChainId } from '@cowprotocol/cow-sdk'
 
 import {
+  DEFAULT_TRADE_DERIVED_STATE,
+  ExtendedTradeRawState,
+  getDefaultTradeRawState,
+  TradeDerivedState,
+} from 'modules/trade'
+import {
   alternativeOrderAtomSetterFactory,
   alternativeOrderReadWriteAtomFactory,
 } from 'modules/trade/state/alternativeOrder'
-import { DEFAULT_TRADE_DERIVED_STATE, TradeDerivedState } from 'modules/trade/types/TradeDerivedState'
-import { ExtendedTradeRawState, getDefaultTradeRawState } from 'modules/trade/types/TradeRawState'
 
 export interface LimitOrdersDerivedState extends TradeDerivedState {
   readonly isUnlocked: boolean
