+++ apps/cowswap-frontend/src/modules/limitOrders/updaters/AlternativeLimitOrderUpdater.ts
@@ -7,22 +7,18 @@ import { useENS } from '@cowprotocol/ens'
 import { useWalletInfo } from '@cowprotocol/wallet'
 import { Price } from '@uniswap/sdk-core'
 
-import {
-  limitOrdersDerivedStateAtom,
-  LimitOrdersSettingsState,
-  updateLimitOrdersSettingsAtom,
-} from 'modules/limitOrders'
-import { useUpdateLimitOrdersRawState } from 'modules/limitOrders/hooks/useLimitOrdersRawState'
-import { LIMIT_ORDERS_DEADLINES } from 'modules/limitOrders/pure/DeadlineSelector/deadlines'
-import { partiallyFillableOverrideAtom } from 'modules/limitOrders/state/partiallyFillableOverride'
-import { useAlternativeOrder, useHideAlternativeOrderModal } from 'modules/trade/state/alternativeOrder'
+import { DEFAULT_TRADE_DERIVED_STATE, useAlternativeOrder, useHideAlternativeOrderModal } from 'modules/trade'
 
 import { GenericOrder } from 'common/types'
 
-import { DEFAULT_TRADE_DERIVED_STATE } from '../../trade'
 import { useLimitOrdersDerivedState } from '../hooks/useLimitOrdersDerivedState'
+import { useUpdateLimitOrdersRawState } from '../hooks/useLimitOrdersRawState'
 import { useUpdateActiveRate } from '../hooks/useUpdateActiveRate'
+import { LIMIT_ORDERS_DEADLINES } from '../pure/DeadlineSelector/deadlines'
+import { limitOrdersDerivedStateAtom } from '../state/limitOrdersRawStateAtom'
+import { LimitOrdersSettingsState, updateLimitOrdersSettingsAtom } from '../state/limitOrdersSettingsAtom'
 import { updateLimitRateAtom } from '../state/limitRateAtom'
+import { partiallyFillableOverrideAtom } from '../state/partiallyFillableOverride'
 
 export function AlternativeLimitOrderUpdater(): null {
   // Update raw state and related settings once on load
