+++ apps/cowswap-frontend/src/modules/ordersTable/pure/OrdersTable/Content/NoOrders/OrdersTableNoOrdersContent.tsx
@@ -7,7 +7,7 @@ import { useIsSafeViaWc } from '@cowprotocol/wallet'
 import { useLingui } from '@lingui/react/macro'
 
 import { useInjectedWidgetParams } from 'modules/injectedWidget'
-import { useLoadMoreOrders } from 'modules/orders/hooks/useLoadMoreOrders'
+import { useLoadMoreOrders } from 'modules/orders'
 
 import { getTitle, getDescription } from './OrdersTableNoOrdersContent.utils'
