+++ apps/cowswap-frontend/src/modules/tokensList/containers/SelectTokenWidget/hooks/useChainAnalyticsContext.ts
@@ -5,7 +5,7 @@ import { useWalletInfo } from '@cowprotocol/wallet'
 
 import { Field } from 'legacy/state/types'
 
-import { TradeType } from 'modules/trade/types'
+import { TradeType } from 'modules/trade'
 
 import { useChainsToSelect } from '../../../hooks/useChainsToSelect'
 import { useSelectTokenWidgetState } from '../../../hooks/useSelectTokenWidgetState'
