+++ apps/cowswap-frontend/src/modules/tokensList/containers/SelectTokenWidget/internal/slots/ChainSelector.tsx
@@ -8,7 +8,7 @@ import { t } from '@lingui/core/macro'
 
 import { Field } from 'legacy/state/types'
 
-import { TradeType } from 'modules/trade/types'
+import { TradeType } from 'modules/trade'
 
 import { MobileChainSelector } from '../../../../pure/SelectTokenModal/MobileChainSelector'
 import { ChainsToSelectState } from '../../../../types'
