+++ apps/cowswap-frontend/src/modules/tokensList/hooks/useOnSelectChain.ts
@@ -4,7 +4,7 @@ import { ChainInfo } from '@cowprotocol/cow-sdk'
 
 import { Field } from 'legacy/state/types'
 
-import { TradeType } from 'modules/trade/types'
+import { TradeType } from 'modules/trade'
 
 import { useSelectTokenWidgetState } from './useSelectTokenWidgetState'
 import { useUpdateSelectTokenWidgetState } from './useUpdateSelectTokenWidgetState'
