+++ apps/cowswap-frontend/src/modules/tokensList/pure/ChainPanel/index.tsx
@@ -7,7 +7,7 @@ import { Trans } from '@lingui/react/macro'
 
 import { Field } from 'legacy/state/types'
 
-import { TradeType } from 'modules/trade/types'
+import { TradeType } from 'modules/trade'
 
 import { ChainPanelHeader } from './ChainPanelHeader'
 import { filterChainsByQuery, getEmptyStateFlags } from './chainPanelUtils'
