+++ apps/cowswap-frontend/src/modules/trade/containers/NoImpactWarning/index.tsx
@@ -5,13 +5,14 @@ import { useWalletInfo } from '@cowprotocol/wallet'
 
 import { Trans } from '@lingui/react/macro'
 
-import { useTradePriceImpact } from 'modules/trade'
 import { TradeWarning } from 'modules/trade/pure/TradeWarning'
 import { TradeWarningType } from 'modules/trade/pure/TradeWarning/constants'
 import { TradeFormValidation, useGetTradeFormValidation } from 'modules/tradeFormValidation'
 import { useTradeQuote } from 'modules/tradeQuote'
 
 import { noImpactWarningAcceptedAtom } from './useIsNoImpactWarningAccepted'
+
+import { useTradePriceImpact } from '../../hooks/useTradePriceImpact'
 
 const NoImpactWarningMessage = (
   <div>
