+++ apps/cowswap-frontend/src/modules/trade/hooks/useDerivedTradeState.ts
@@ -1,8 +1,7 @@
 import { useAtomValue } from 'jotai'
 
-import { TradeDerivedState } from 'modules/trade/types/TradeDerivedState'
-
 import { derivedTradeStateAtom } from '../state/derivedTradeStateAtom'
+import { TradeDerivedState } from '../types/TradeDerivedState'
 
 export function useDerivedTradeState(): TradeDerivedState | null {
   return useAtomValue(derivedTradeStateAtom)
