+++ apps/cowswap-frontend/src/modules/trade/index.ts
@@ -1,10 +1,26 @@
+// Export types first to avoid circular dependency (getDefaultTradeRawState etc. used by yield, limitOrders, etc.)
+export * from './types'
+export * from './const/common'
+export * from './state/tradeTypeAtom'
+export * from './state/derivedTradeStateAtom'
+export * from './state/isWrapOrUnwrapAtom'
+export * from './state/isEoaEthFlowAtom'
+export { useAlternativeOrder, useHideAlternativeOrderModal } from './state/alternativeOrder/hooks'
+export * from './utils/parameterizeTradeRoute'
+export * from './utils/parameterizeTradeSearch'
+export * from './utils/getCrossChainReceiveAmountInfo'
+export * from './utils/getTotalCosts'
+export * from './utils/getOrderTypeReceiveAmounts'
+export * from './utils/tradeFlowAnalytics'
+export * from './utils/logger'
+export { addPendingOrderStep } from './utils/addPendingOrderStep'
+export * from './const/tradeUrl'
 export * from './containers/TradeWidget'
 export * from './containers/TradeConfirmModal'
 export * from './containers/TradeWidgetLinks'
 export * from './containers/TradeFeesAndCosts'
 export * from './containers/TradeTotalCostsDetails'
 export * from './containers/TradeBasicConfirmDetails'
-export * from './const/common'
 export * from './pure/TradeConfirmation'
 export * from './hooks/useTradeConfirmActions'
 export * from './hooks/useConfirmTradeWithRwaCheck'
@@ -14,7 +30,6 @@ export * from './hooks/setupTradeState/useSetupTradeState'
 export * from './hooks/useWrapNativeFlow'
 export * from './hooks/useDerivedTradeState'
 export * from './hooks/useSetupTradeAmountsFromUrl'
-export * from './types/TradeDerivedState'
 export * from './hooks/useTradeNavigate'
 export * from './hooks/useTradeState'
 export * from './hooks/useIsNativeInOrOut'
@@ -52,16 +67,6 @@ export type { AmountsToSign } from './hooks/useAmountsToSignFromQuote'
 export { useGetTradeUrlParams } from './hooks/useGetTradeUrlParams'
 export * from './containers/TradeWidget/types'
 export { useIsNoImpactWarningAccepted } from './containers/NoImpactWarning/useIsNoImpactWarningAccepted'
-export * from './utils/getCrossChainReceiveAmountInfo'
-export * from './utils/getTotalCosts'
-export * from './utils/getOrderTypeReceiveAmounts'
-export * from './utils/parameterizeTradeRoute'
-export * from './utils/tradeFlowAnalytics'
-export * from './utils/parameterizeTradeSearch'
-export * from './state/tradeTypeAtom'
-export * from './state/derivedTradeStateAtom'
-export * from './state/isWrapOrUnwrapAtom'
-export * from './state/isEoaEthFlowAtom'
 export * from './pure/RecipientRow'
 export * from './pure/ReceiveAmountTitle'
 export * from './pure/FreeFeeRow'
@@ -74,4 +79,3 @@ export * from './pure/ReviewOrderModalAmountRow'
 export * from './pure/ConfirmDetailsItem'
 export { CompatibilityIssuesWarning } from './pure/CompatibilityIssuesWarning'
 export * from './pure/Row/styled'
-export * from './types'
