+++ apps/cowswap-frontend/src/pages/Swap/index.tsx
@@ -8,8 +8,7 @@ import { Navigate, useLocation, useParams } from 'react-router'
 
 import { PageTitle } from 'modules/application/containers/PageTitle'
 import { SwapUpdaters, SwapWidget } from 'modules/swap'
-import { getDefaultTradeRawState } from 'modules/trade/types/TradeRawState'
-import { parameterizeTradeRoute } from 'modules/trade/utils/parameterizeTradeRoute'
+import { getDefaultTradeRawState, parameterizeTradeRoute } from 'modules/trade'
 
 import { Routes } from 'common/constants/routes'
