+++ apps/cowswap-frontend/src/serviceWorker/document.test.ts
@@ -3,7 +3,6 @@ import { getCacheKeyForURL as getCacheKeyForURLMock, matchPrecache as matchPreca
 
 import { CachedDocument, handleDocument, matchDocument } from './document'
 
-jest.mock('workbox-navigation-preload', () => ({ enable: jest.fn() }))
 jest.mock('workbox-precaching', () => ({
   getCacheKeyForURL: jest.fn(),
   matchPrecache: jest.fn(),
